added plan liquidity denom
